Move ErrorActions under actions folder
