+++ client/src/error/ErrorContainer.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux'
 import ErrorComponent from './ErrorComponent'
 import { push, goBack } from 'react-router-redux'
-import { clearErrors } from './ErrorActions'
+import { clearErrors } from '../actions/ErrorActions'
 import { clearSearch } from '../actions/SearchActions'
 
 const mapStateToProps = (state) => {
