+++ client/test/actions/searchActionSpec.js
@@ -2,7 +2,7 @@ import '../specHelper'
 import * as module from '../../src/actions/SearchActions'
 import { LOADING_SHOW, LOADING_HIDE } from '../../src/loading/LoadingActions'
 import { FACETS_RECEIVED, CLEAR_FACETS } from '../../src/search/facet/FacetActions'
-import { SET_ERRORS } from '../../src/error/ErrorActions'
+import { SET_ERRORS } from '../../src/actions/ErrorActions'
 import configureMockStore from 'redux-mock-store'
 import thunk from 'redux-thunk'
 import Immutable from 'seamless-immutable'
